[feat] Add trash feature
Soft-delete subsystem: unlinked entries are moved into a per-fs
.trash directory bucketed by hour, kept until GC reclaims them
past retention, and restorable via a CLI tool. Modeled on JuiceFS
trash semantics.

MDS:
- FileSystem routes Unlink/RmDir/BatchUnLink through the trash
  pipeline; Rename mv-out and root-only manual cleanup are
  supported. Symlinks go through trash like regular files.
- Trash entry name encodes origin_parent; cleanup walks ancestors
  via UpdateDirUsage (no stored quota_chain).
- Create-side guards (Mk*/Link/Symlink/Rename) forbid create-in-
  trash and create-in-trashed-dir via a parent_memo_ -> cache -> KV
  cascade in CheckCreateInTrash, closing the smuggle-via-mv-out
  hole that JuiceFS's pattr.Parent>TrashInode check covers.
- ApplyDeleteQuota helper unifies the 4-case delete quota matrix
  (trash-cleanup / soft / deferred / permanent) across the three
  delete ops.
- Synthetic .trash inode attr pinned to fs.create_time_s so attr
  caches do not churn and stat() is idempotent.
- ParentHashMDSRouter falls back to a random MDS when parent is
  kTrashInodeId, avoiding the hotspot on .trash listing and
  hour-bucket lookups; receivers read directly from KV.
- SubTrashCache (per-fs bucket cache) is encapsulated behind
  BthreadRWLock - read-mostly: Lookup on every delete, Record only
  on hour rollover.
- Background GC walks .trash hourly; restore tool
  (src/mds/client/trash_restore.{h,cc}) reverses an entry back to
  its origin parent.

Client:
- Surface .trash as a virtual directory under the mount root
  (kTrashIno, kTrashDirName); IsInternalNode/IsInternalName
  recognize it.
- Plumb FUSE caller uid through Context so the Rename trash-out
  guard (old_parent==kTrashIno && uid!=0) fires client-side.
- Pre-flight Mk*/Link/Symlink/Rename/Unlink/RmDir trash guards
  mirror MDS rejections to fail fast; root manual cleanup falls
  through to MDS, which authorizes.

Proto:
- FsInfo.create_time_s, Context.uid; trash bucket children reuse
  the Dentry codec with parent-aware decode branches (no wrapper
  message).

Tests:
- common/test_trash.cc: trash inode boundaries, entry-name and
  bucket-name codec round-trips.
- filesystem/test_trash.cc: end-to-end Unlink->.trash->Rename mv-out,
  RmDir, BatchUnLink incl. symlinks, hour-bucket lifecycle, GC,
  restore-tool integration; regressions for synthetic-root op
  rejects, multi-partition same-name rename, BatchGet dedup on
  mv-out, root-can-rename-bucket-out, and
  CannotCreateInsideTrashedDirectory.
- client/vfs/metasystem/test_mds_router_trash.cc:
  parent==kTrashInodeId routing fallback.

+++ src/mds/common/runnable.cc
@@ -306,7 +306,10 @@ bool ExecqWorkerSet::ExecuteHash(int64_t id, TaskRunnablePtr task) {
     return false;
   }
 
-  auto ret = workers_[id % WorkerNum()]->Execute(task);
+  // Cast to uint64_t before modulo: callers pass Ino (uint64_t), and sub-trash
+  // bucket inos (kTrashInodeId + offset) can exceed INT64_MAX, producing a
+  // negative id and thus workers_[negative] OOB.
+  auto ret = workers_[static_cast<uint64_t>(id) % WorkerNum()]->Execute(task);
   if (ret) {
     IncPendingTaskCount();
     IncTotalTaskCount();
